requirements.txt
imgurpython==1.1.7
Pillow==8.1.0
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.2
